pyproject.toml
[project]
name = "historical-debate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[extensions] >=1.26", 
    "loguru>=0.7.3",
]
